refactor(p2-shim): separate in-memory filesystem adapter
