+++ packages/preview2-shim/src/browser/in-memory-filesystem.ts
@@ -1,0 +1,11 @@
+import type { BrowserFilesystemAdapter, FileData } from "./filesystem.js";
+
+/** Explicit ephemeral storage adapter for browser applications and tests. */
+export class InMemoryFilesystemAdapter implements BrowserFilesystemAdapter<FileData> {
+    getRoot(capability: FileData): FileData {
+        if (!capability.dir) {
+            throw new TypeError("an in-memory preopen root must be a directory");
+        }
+        return capability;
+    }
+}
